make receivedItemMessage contain graph::element directly

+++ src/archipelago/consts.rs
@@ -1,2 +1,2 @@
-pub(super) const ELEMENT_ID_OFFSET: usize = 100;
+pub(super) const ELEMENT_ID_OFFSET: u64 = 100;
 pub(super) const START_ITEMS: u64 = 4;
